requirements.txt
fastapi==0.101.0
gevent==23.7.0
tritonclient[http]==2.3.0
nemo-toolkit[asr]==1.18.1
pydantic==2.1.1
pydub==0.25.1
numpy==1.23
Requests==2.31.0
torch==2.0.1
torchmetrics==0.11.4
uvicorn[standard]==0.15.0
hydra-core==1.2.0
loguru